fix : HistoryRecoder 로직 분리
