--- src/main/java/com/workhub/global/clientInfo/ClientInfo.java
@@ -1,12 +1,0 @@
-package com.workhub.global.clientInfo;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Deprecated
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ClientInfo {
-}
