--- src/main/java/com/workhub/global/clientInfo/ClientInfoArgumentResolver.java
@@ -1,31 +1,0 @@
-package com.workhub.global.clientInfo;
-
-import com.workhub.global.util.RequestUtil;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Deprecated
-@Component
-public class ClientInfoArgumentResolver implements HandlerMethodArgumentResolver {
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(ClientInfo.class);
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-
-        HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-
-        String ip = RequestUtil.getClientIp(request);
-        String userAgent = RequestUtil.getUserAgent(request);
-
-        return new ClientInfoDto(ip,userAgent);
-    }
-}
