+++ src/main/java/com/workhub/global/history/HistoryCreator.java
@@ -1,0 +1,10 @@
+package com.workhub.global.history;
+
+import com.workhub.global.entity.ActionType;
+import com.workhub.global.entity.BaseHistoryEntity;
+
+@FunctionalInterface
+public interface HistoryCreator {
+    BaseHistoryEntity create(Long targetId, ActionType actionType, String beforeData,
+                             Long creator);
+}
